% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_class.R
\name{Bin_mono}
\alias{Bin_mono}
\title{Set monotoncity of performance}
\arguments{
\item{m}{the monotonic relationship to enforce.}
}
\description{
\code{mono} calls the \code{bin} function with the requested
monotoncity. The variable is discretized while enforcing the monotonicity.
The possible values are:
\itemize{
 \item{0 }{ No monotoncity enforced - the default.}
 \item{1 }{ Increasing monotoncically with the \code{y}}
 \item{-1 } {Decreasing monotoncically with the \code{y}}
 \item{2 }{ Either increasing or decreasing montonically with the \code{y}}
}
}
