% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_class.R
\name{Continuous_factorize}
\alias{Continuous_factorize}
\title{Preprocess transform object for summarization}
\arguments{
\item{newdata}{Numeric vector on which to apply the transformation. Defaults
to the \code{x} field of the Continuous object}
}
\value{
\code{list} with two fields. See details.
}
\description{
Preprocess transform object for summarization
}
\details{
\code{factorize} returns a list with two fields:
\itemize{
 \item{factor }{ Factor with the bin labels applied to \code{x}}
 \item{types }{ list of logical vectors for missing, exception, and normal}
}
}
