% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\name{Discrete_collapse}
\alias{Discrete_collapse}
\alias{Discrete_expand}
\alias{Discrete_factorize}
\alias{Discrete_predict}
\alias{Discrete_gen_code_sas}
\title{Collapse levels of a Discrete bin object}
\arguments{
\item{i}{numeric vector of bin levels to collapse. Do not have to be
adjacent.}

\item{i}{numeric vector of length 1 indiicating bin level to expand.}

\item{newdata}{Factor vector on which to apply the transformation. Defaults
to the \code{x} field of the Discrete object}

\item{newdata}{Factor vector to apply performance substition. Defaults to
data used to create the Discrete object.}

\item{pfx}{character prefix to prepend to variable names}

\item{coef}{numeric coefficient to multiply performance values by. Passed in
by the Scorecard model object. Defaults to 1.}

\item{method}{method used for calculating the reference level for adverse
action codes. Three possible choices:
\itemize{
 \item{"min" }{Calculate difference from minimum of perf values - default}
 \item{"max" }{Calculate difference from maximum of perf values}
 \item{"neutral" }{Calculate difference from zero}#'
 }}

\item{i}{numeric value enumerating the variables. Passed in from other code.}
}
\value{
modifies the transform object in place.

modifies the transform object in place.

\code{list} with two fields. See details.

numeric variable with bin performance values substituted for
the inputs.

a character vector of SAS code
}
\description{
Collapse levels of a Discrete bin object

Expand a level of a Discrete bin into multiple new levels

Preprocess transform object for summarization

Substitute summarized performance values for numeric inputs

generate SAS code representing the transformation from input
numeric values to the substituted performance values. Also generates code
calculating difference from min/max/neutral and adverse action code
assignments.
}
\details{
All of the collapsed levels will be expanded.

\code{factorize} returns a list with two fields:
\itemize{
 \item{factor }{ Factor with the bin labels applied to \code{x}}
 \item{types }{ list of logical vectors for missing, exception, and normal}
}
}
