% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_class.R
\name{Discrete_predict}
\alias{Discrete_predict}
\title{Substitute summarized performance values for numeric inputs}
\arguments{
\item{newdata}{Factor vector to apply performance substition. Defaults to
data used to create the Discrete object.}
}
\value{
numeric variable with bin performance values substituted for
the inputs.
}
\description{
Substitute summarized performance values for numeric inputs
}
