% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_class.R
\docType{class}
\name{Scorecard_Class}
\alias{Scorecard_Class}
\alias{Scorecard}
\alias{Scorecard_select}
\title{Select a fitted model and load the Transforms}
\description{
Scorecard class that wraps a data.frame and prepares it for
scorecard modeling.

Select searches the model fitting history for the requested
model. It then loads the associated transforms for the model variables into
the current Scorevard object. Additionally it loads the dropped and inmodel
vectors at the time of the fit.
}
\section{Fields}{

\describe{
\item{\code{seed}}{saved random seed that is based on  \code{\link{Sys.time}}.}

\item{\code{models}}{list of fitted models for this Scorecard}

\item{\code{selected_model}}{name of last  selected that was loaded into the
transforms}

\item{\code{inmodel}}{character vector of variable names that are in the selected
model}
}}


