% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_class.R
\docType{class}
\name{Transform_Class}
\alias{Transform_Class}
\alias{neutralize}
\alias{neutralize_}
\alias{set_equal}
\alias{set_equal_}
\alias{update_transform}
\title{Neutralize selected levels of Transform setting substitution to zero}
\usage{
neutralize_(tf, i)

set_equal_(tf, v1, v2)

update_transform(tf, result)
}
\arguments{
\item{tf}{Transform object}

\item{i}{numeric vector of levels to neutralize}

\item{v1}{level to override}

\item{v2}{value with which to override v1's substitution}

\item{result}{list of summarized performane values for the updated Bin object}

\item{tf}{Transform object}

\item{tf}{Transform object}
}
\value{
new Transform object with updated overrides

new Transform object with updated overrides

new Transform object with updated subst, nas, exceptions, and repr
}
\description{
Transform object class definition. The Transform object controls
how an independent variable is binned and manages missing values, exceptions,
overrides, and weight-of-evidence substitution.

Neutralize selected levels of Transform setting substitution to zero

Set substitution of one level equal to substitution of another

Update Transform object with new information after Bin operations
}
\section{Slots}{

\describe{
\item{\code{tf}}{the mapping of input values to bin levels}

\item{\code{subst}}{the WoE substitution for the mapped values}

\item{\code{exceptions}}{numeric vector of exception values}

\item{\code{nas}}{WoE value for missing value sof \code{x}}

\item{\code{overrides}}{named vector of bin levels to overriden substituion values}

\item{\code{repr}}{cached, tabular representation of the bin object for quick display}
}}

